Create Fibonacci_Sequence_Generator.py
This code defines a function, fibonacci, that takes a parameter n, representing the length of the Fibonacci sequence to be generated. The code first creates an empty list, fib_list, to store the Fibonacci sequence. Then, add 0 and 1 to the list as the first two numbers. Next, use a loop to generate the remaining numbers starting with the third number, each of which is the sum of the first two, and then add it to the list. Finally, the generated Fibonacci sequence is returned.



In the main code, the user is asked to enter the length of the fibonacci sequence to be generated and call the Fibonacci function to generate the sequence. Finally, print the generated Fibonacci sequence.